leaveapp/requirements.txt
asgiref==3.8.1
certifi==2023.7.22
charset-normalizer==3.3.2
distlib==0.3.7
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
idna==3.9
numpy==1.26.1
pipenv==2023.10.24
platformdirs==3.11.0
psycopg2==2.9.9
PyJWT==2.9.0
requests==2.32.3
setuptools==68.2.2
sqlparse==0.4.4
stripe==10.11.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
virtualenv==20.24.6
